+++ libs/ui-lib/lib/ocm/components/wizard/utils/wizardTransition.ts
@@ -382,27 +382,3 @@ export const canNextNetwork = ({ cluster }: TransitionProps): boolean =>
 export const isStaticIpStep = (stepId: ClusterWizardStepsType | Day2WizardStepsType) => {
   return stepId.startsWith('static-ip');
 };
-
-export type WizardNavEntry = {
-  stepId: ClusterWizardStepsType;
-  visualNumber: number;
-  isStaticIpFormGroup: boolean;
-};
-
-export const getWizardNavEntries = (wizardStepIds: ClusterWizardStepsType[]): WizardNavEntry[] => {
-  const entries: WizardNavEntry[] = [];
-  let i = 0;
-  let visualNumber = 1;
-  while (i < wizardStepIds.length) {
-    const stepId = wizardStepIds[i];
-    if (stepId === 'static-ip-network-wide-configurations') {
-      entries.push({ stepId, visualNumber, isStaticIpFormGroup: true });
-      i += 2;
-    } else {
-      entries.push({ stepId, visualNumber, isStaticIpFormGroup: false });
-      i += 1;
-    }
-    visualNumber += 1;
-  }
-  return entries;
-};
